status al 16/12

+++ src/main/java/com/springboot/currentAccount/dto/CurrentAccountEnterDto.java
@@ -8,6 +8,6 @@ public class CurrentAccountEnterDto {
 	private String numberAccount;
 	private String tea;
 	private String state;
-	private int balance;
+	private Double balance;
 	private EnterpriseDto holders;
 }
